% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{climate_locations}
\alias{climate_locations}
\title{Display available locations}
\usage{
climate_locations(type = "all")
}
\arguments{
\item{type}{character, \code{"all"}, \code{"region"} or \code{"point"}.}
}
\value{
a data frame.
}
\description{
Display a data frame of available locations and their respective location groups/sets.
}
\examples{
climate_locations()
}
