% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{climate_regions}
\alias{climate_regions}
\title{Climate regions}
\format{A data frame with 86 rows and 2 columns giving the region name and the set/group it belongs to.}
\usage{
climate_regions
}
\description{
A data frame of available climate regions and their respective region groups for use in \code{\link{climdata}}.
}
\keyword{datasets}
