% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapclim.R
\name{climdata}
\alias{climdata}
\title{Obtain SNAP climate data sets}
\usage{
climdata(id, area, time_scale = "monthly", set = NULL)
}
\arguments{
\item{id}{character, data set ID. See \code{\link{collections}} for available data sets.}

\item{area}{character, aerial unit/region or point location of interest. See \code{\link{climate_regions}}.}

\item{time_scale}{character, \code{"monthly"}, \code{"seasonal"} or \code{"annual"}.}

\item{set}{character or \code{NULL}, The set/group that \code{area} belongs to. Can be ignored,
but should be provided in rare cases where \code{area} name is not unique (a warning is thrown). See \code{\link{climate_regions}}.}
}
\value{
a data frame.
}
\description{
Download climate data sets from SNAP.
}
\details{
This function downloads climate data sets from Scenarios Network for Alaska and Arctic Planning based on the data set \code{id}
and specification of the region or point location of interest.
}
\examples{
climdata("ar5stats", "AK-CAN")
}
