% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapclim.R
\name{collections}
\alias{collections}
\title{Basic metadata for all data sets available via snapclim}
\usage{
collections()
}
\value{
a data frame.
}
\description{
This function returns a data frame with basic meta data for all of the curated collections of SNAP climate data
available via \code{snapclim}.
This includes ID, short description, and spatial and temporal information.
}
\examples{
collections()
}
