% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{location_sets}
\alias{location_sets}
\title{List the available location sets}
\usage{
location_sets(type)
}
\arguments{
\item{type}{character, \code{"region"} or \code{"point"}.}
}
\value{
a character vector.
}
\description{
List the available location sets/groups for regional or point location data.
}
\examples{
location_sets("region")
location_sets("point")
}
